refactor(execd): trim OSEP-0018 tests, comments and dead code
Test dedup:
- init_container.sh: drop exit-code propagation and runtime-SIGTERM
  cases and the floor assertions, all covered by the Python e2e suites
  (test_execd_init_e2e.py, test_execd_hardening_e2e.py); keep the
  smoke-layer contract the e2e does not reach (PID 1 handoff, orphan
  reaping, signal shield, subreaper degradation, entrypoint env
  inheritance incl. JUPYTER_TOKEN preservation)
- test_execd_init_e2e.py: drop test_fork_heavy_keeps_process_table_bounded
  (subsumed by the sustained fork-heavy test)
- test_execd_hardening_e2e.py: drop the duplicate
  test_workload_cannot_read_execd_environ (same pin exists in the init e2e)

Comment cleanup:
- condense verbose OSEP-0018 comments in execd code and tests; fix the
  duplicated hardenCmd doc comment in hardening_linux.go

Dead code:
- launcher.c: remove the caps_dropped counter (written, never read)
- remove initmode_barrier_windows.go: waitManagedWithBarrier has no
  callers on Windows (the only call site is in a !windows file)
- split initmode_other.go: the managedProcess fallback is Windows-dead
  and moves to initmode_other_unix.go (!linux && !windows)
- exit_error.go: restrict exitCoder to !windows (no Windows callers)

Verified: go build + go vet on linux/windows/darwin, staticcheck U1000
clean for all OSEP-0018 code across build variants.
