+++ components/execd/pkg/runtime/exit_error.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 // Copyright 2026 Alibaba Group Holding Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
