--- components/execd/pkg/runtime/initmode_barrier_windows.go
@@ -1,25 +1,0 @@
-//go:build windows
-
-// Copyright 2026 Alibaba Group Holding Ltd.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package runtime
-
-// waitManagedWithBarrier mirrors waitCommandWithExitBarrier's semantics on
-// platforms where that helper does not exist.
-func waitManagedWithBarrier(mp *managedProcess, mark func(error)) error {
-	err := mp.Wait()
-	mark(nil)
-	return err
-}
